Prevent spawning multiple windows (#109)
Refs #107. 

Prevent spawning multiple windows by starting a Unix socket listener as
an Iced subscription. Before launching a new instance, check if that
socket can be connected to, and refuse opening if so.

This is intentionally lax: if we ended up using a named file and it has
persisted a crash of a previous app instance, we want to still launch a
new window. If the listener fails to start, we log that error and keep
running, because this functionality is non-essential.